chore: deprecate selector-util matchSelectorTarget
- move to deprecated folder
- mark as deprecated on index export
- kept tests until removed (moved to proper folder)
